+++ src/main/java/nextstep/subway/applicaion/dto/station/StationResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.applicaion.dto.station;
 
 public class StationResponse {
     private Long id;
