+++ src/main/java/nextstep/subway/domain/line/LineRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.line;
 
 import org.springframework.data.jpa.repository.JpaRepository;
